+++ docs/api/changelog.rst
@@ -22,6 +22,9 @@ Version 0.6
   The resolver can be overridden (disabling the reconfiguration magic) using
   :func:`aioxmpp.network.set_resolver`.
 
+* :meth:`aioxmpp.muc.Room.on_message` now receives a proper `occupant` argument
+  if occupant data is available when the message is received.
+
 Version 0.5
 ===========